Make string concatenation operator use consistent.
The code was using two different formats: (1) the string concatenation
operator is immediately adjacent to string literals, and (2) the string
concatenation operator is surrounded by spaces.  The first was more
common in the code, but the second is the format specified by the Drupal
Coding Standards.  The second is also the format expected by Coder
Review.  (Coder Review used to expect the first.)  Therefore, chose to
surround concatenation operator with spaces, though both formats seem
fine to me.