Improve expired password redirection.
Prior to this change, upon password expiration, we use drupal_goto() to
send the user to their destination page, which causes them to be
redirected to the password change page by password_policy_init().  Now
we use drupal_goto() to send the user directly to the password change
page, which saves a page request.