Change $value parameter of field assertions from '' to NULL.
The '' $value parameter to assertNoFieldByName() means to assert that
there is no field with the empty string value. This is probably not what
was intended by the author of these tests. It looks like the intent was
only to assert that the fields did not exist. We change assertField*()
calls to use NULL for parity.