Remove completed @TODO.
The author of 6d63b3a8 added a to-do to "findout why we cannot use
arguments here".  It seems like they were referring to the subsequent
line.  That line has a leftJoin() call which uses "%s" for substitutions.
However, "%s" is not supported for substitutions in Drupal 7.  I suspect
the author did not know this and it is what prompted the comment.